# Conflicts:
#	src/main/java/edu/uga/miage/m1/polygons/gui/JDrawingFrame.java
#	src/main/java/edu/uga/miage/m1/polygons/gui/persistence/JSONSaver.java
#	src/main/java/edu/uga/miage/m1/polygons/gui/persistence/XMLSaver.java

+++ src/main/java/edu/uga/miage/m1/polygons/gui/shapes/Square.java
@@ -51,7 +51,7 @@ public class Square implements SimpleShape, Visitable {
      */
     public void draw(Graphics2D g2) {
         g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-        GradientPaint gradient = new GradientPaint(m_x, m_y, Color.BLUE, m_x + 50, m_y, Color.WHITE);
+        GradientPaint gradient = new GradientPaint((float) m_x, (float) m_y, Color.BLUE, (float) m_x + 50, m_y, Color.WHITE);
         g2.setPaint(gradient);
         g2.fill(new Rectangle2D.Double(m_x, m_y, 50, 50));
         BasicStroke wideStroke = new BasicStroke(2.0f);
